requirements.txt
streamlit
opencv-python
numpy
Pillow
cvlib
tensorflow
keras